# topgun-services/docs/swagger.yaml
definitions:
  models.Camera:
    properties:
      id:
        type: string
      institute:
        type: string
      location:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  models.Detect:
    properties:
      camera:
        $ref: '#/definitions/models.Camera'
      camera_id:
        type: string
      id:
        type: integer
      path:
        type: string
    type: object
...
info:
  contact: {}
  title: KKU GS ADMISSION SERVICES API
paths:
...
  /api/v1/detect/:
    get:
      consumes:
      - application/json
      description: Get a list of detects with pagination and filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search filter
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetDetects
      tags:
      - Detect
    post:
      consumes:
      - multipart/form-data
      description: Create a new detect with file upload
      parameters:
      - description: Camera ID (UUID)
        in: formData
        name: camera_id
        required: true
        type: string
      - description: Objects JSON array
        in: formData
        name: objects
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: CreateDetect
      tags:
      - Detect
  /api/v1/detect/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a detect by ID
      parameters:
      - description: Detect ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: DeleteDetect
      tags:
      - Detect
    get:
      consumes:
      - application/json
      description: Get a detect by ID
      parameters:
      - description: Detect ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetDetect
      tags:
      - Detect
    put:
      consumes:
      - application/json
      description: Update an existing detect
      parameters:
      - description: Detect ID
        in: path
        name: id
        required: true
        type: string
      - description: Detect object
        in: body
        name: detect
        required: true
        schema:
          $ref: '#/definitions/models.Detect'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: UpdateDetect
      tags:
      - Detect
  /api/v1/detect/{id}/file:
    get:
      description: Download the file associated with a detect by ID
      parameters:
      - description: Detect ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetDetectFile
      tags:
      - Detect
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"